# server/src/swagger/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ApiResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					}
				},
				"required": [
					"message",
					"data",
					"statusCode",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Add Bearer prefix on jwt token"
			}
		}
	},
	"info": {
		"title": "GDSC JIIT-128 API Documentaion",
		"version": "1.1.1",
		"description": "Routes of main site of gdsc jiit-128 api documentation",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/signup-user": {
			"post": {
				"operationId": "SignupUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "signup new user",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"role": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"mobileNumber": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"role",
									"email",
									"mobileNumber",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "Login user",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"role": {
										"type": "string"
									},
									"emailOrMobile": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"role",
									"emailOrMobile"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/verify-email/{token}": {
			"get": {
				"operationId": "VerifyEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "pass token and get your email varified",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/send-email-otp": {
			"post": {
				"operationId": "SendEmailOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "send email otp",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/verify-email-otp": {
			"post": {
				"operationId": "VerifyEmailOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "verify email otp",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"otp": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"otp",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/change-password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "change password",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"confirm_password": {
										"type": "string"
									},
									"new_password": {
										"type": "string"
									}
								},
								"required": [
									"confirm_password",
									"new_password"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/social-network-authentication": {
			"post": {
				"operationId": "SocialNetworkAuthentication",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "signup or login using social networking  sites",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"avatar": {
										"type": "string"
									},
									"role": {
										"type": "string"
									},
									"socialId": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"avatar",
									"role",
									"socialId",
									"email",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000/api/v1",
			"description": "Local environment"
		},
		{
			"url": "https://something.com/api/v1",
			"description": "Some environment"
		}
	]
}